Updates and Bug Fixes
•    "Search For Mods" tab has a new look, more refined searching, and
now displays mod information without opening a new window. Pictures have
been updated but here is one as well.
•    Fixed an issue where implementing specific types of mods (VFX were
the biggest offender) did not install properly [Thanks to Flowey for the
catch].
•    Auto updates are now a thing! This will only be available to all
users of this version and up. This is a big step in ensuring a smooth
transition into a full release so if you find any issues after an
update, please leave feedback in the feedback tab, here in this thread,
or PM me.
•    In the case of an issue with auto updating, a debug button (located
in the feedback tab) will allow me to give user inputs to hopefully
ensure updates occur as they should.
•    Added a link to this thread / guide. I received many random
feedback logs that consisted of users not knowing what to do.
•    Changed the Live Folder examples into a tool tip. It was in an
awkward position in my opinion and the old link is now replaced by the
link to this guide.
•    Rearranged some things in the initial setup.